dev backend

+++ orderSystem-backend/src/main/java/com/dianchong/ordersystem/entity/DcUser.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -12,7 +13,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 public class DcUser {
-    private Integer userId;
+    private BigDecimal userId;
     private String username;
     private String userAccount;
     private String password;
